Move generic format utilities to `terminalformat` package
