+++ src/lib/jotaen/terminalformat/init_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package tf
+package terminalformat
 
 import (
 	"os"
