+++ src/lib/jotaen/terminalformat/table.go
@@ -1,0 +1,8 @@
+package terminalformat
+
+type Table struct {
+}
+
+func (t *Table) cell(int, string) {
+
+}
